# rewrite loop rint into HIGHWAY