--- a/numpy/_core/meson.build
+++ b/numpy/_core/meson.build
@@ -106,7 +106,10 @@ if use_highway
   highway_lib = static_library('highway',
     [
       # required for hwy::Abort symbol
-      'src/highway/hwy/abort.cc'
+      'src/highway/hwy/abort.cc',
+      'src/highway/hwy/per_target.cc',
+      'src/highway/hwy/targets.cc',
+      'src/highway/hwy/print.cc'
     ],
     cpp_args: '-DTOOLCHAIN_MISS_ASM_HWCAP_H',
     include_directories: ['src/highway'],
@@ -972,12 +975,12 @@ foreach gen_mtargets : [
   ],
   [
     'loops_unary_fp.dispatch.h',
-    src_file.process('src/umath/loops_unary_fp.dispatch.c.src'),
+    'src/umath/loops_unary_fp.dispatch.cpp',
     [
-      SSE41, SSE2,
+      SSE42, SSE2,
       VSX2,
       ASIMD, NEON,
-      VXE, VX
+      VXE, VX, RVV
     ]
   ],
   [