@Mousius 2024-04-30
Hi @luyahan,

Have you measured if this changes/increases performance? Would be good to see some benchmarks 😸 

Just a process point, I assume https://github.com/google/highway/pull/2116 needs to be released before this can be merged?