@luyahan 2024-05-07
> Hi @luyahan,
> 
> Have you measured if this changes/increases performance? Would be good to see some benchmarks 😸
> 
> Just a process point, I assume [google/highway#2116](https://github.com/google/highway/pull/2116) needs to be released before this can be merged?

| Change   | Before [a83d4690]   | After [efd9879e] <rint-hwy>   |   Ratio | Benchmark (Parameter)                                                      |
|----------|----------------------------------------|-------------------------------|---------|----------------------------------------------------------------------------|
| +        | 1.75±0.03μs                            | 5.27±0μs                      |    3.01 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'floor'>, 1, 1, 'f')         |
| +        | 1.74±0.01μs                            | 5.24±0μs                      |    3.01 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'trunc'>, 1, 1, 'f')         |
| +        | 1.77±0μs                               | 5.25±0.01μs                   |    2.96 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'absolute'>, 1, 1, 'f')      |
| +        | 1.77±0.01μs                            | 5.23±0.01μs                   |    2.95 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'square'>, 1, 1, 'f')        |
| +        | 1.93±0μs                               | 5.26±0.01μs                   |    2.73 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rint'>, 1, 1, 'f')          |
| +        | 2.18±0.01μs                            | 5.24±0μs                      |    2.4  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'ceil'>, 1, 1, 'f')          |
| +        | 2.87±0.01μs                            | 5.24±0.01μs                   |    1.83 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sqrt'>, 1, 1, 'f')          |
| +        | 2.88±0.01μs                            | 5.25±0μs                      |    1.82 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'reciprocal'>, 1, 1, 'f')    |
| +        | 3.66±0.01μs                            | 6.10±0.01μs                   |    1.67 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sqrt'>, 4, 1, 'f')          |
| +        | 4.14±0.02μs                            | 6.86±0.01μs                   |    1.66 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 4, 1, 'd')    |
| +        | 4.13±0.03μs                            | 6.83±0.01μs                   |    1.65 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 1, 1, 'd')    |
| +        | 3.66±0μs                               | 5.99±0.01μs                   |    1.64 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'floor'>, 4, 1, 'f')         |
| +        | 3.72±0μs                               | 6.09±0.01μs                   |    1.64 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'reciprocal'>, 4, 1, 'f')    |
| +        | 3.66±0μs                               | 5.96±0μs                      |    1.63 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'ceil'>, 4, 1, 'f')          |
| +        | 3.66±0μs                               | 5.97±0.01μs                   |    1.63 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rint'>, 4, 1, 'f')          |
| +        | 3.66±0.01μs                            | 5.97±0.01μs                   |    1.63 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'trunc'>, 4, 1, 'f')         |
| +        | 4.49±0.02μs                            | 6.82±0.01μs                   |    1.52 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 1, 1, 'e')    |
| +        | 4.49±0.02μs                            | 6.83±0.03μs                   |    1.52 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 4, 1, 'e')    |
| +        | 3.69±0.01μs                            | 5.49±0μs                      |    1.49 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'square'>, 4, 1, 'f')        |
| +        | 4.61±0.07μs                            | 6.84±0.01μs                   |    1.48 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 4, 2, 'd')    |
| +        | 4.67±0.01μs                            | 6.84±0.01μs                   |    1.47 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 1, 2, 'e')    |
| +        | 4.66±0.01μs                            | 6.84±0.01μs                   |    1.47 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 4, 2, 'e')    |
| +        | 4.76±0.01μs                            | 6.84±0.01μs                   |    1.44 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc '_ones_like'>, 1, 2, 'd')    |
| +        | 3.70±0μs                               | 5.34±0.01μs                   |    1.44 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'absolute'>, 4, 1, 'f')      |
| +        | 5.29±0.01μs                            | 6.91±0.05μs                   |    1.31 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 4, 1, 'e')      |
| +        | 5.26±0.01μs                            | 6.85±0.01μs                   |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 1, 1, 'd') |
| +        | 5.26±0.01μs                            | 6.84±0.01μs                   |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 1, 1, 'e') |
| +        | 5.27±0.01μs                            | 6.84±0.02μs                   |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 1, 2, 'e') |
| +        | 5.27±0.01μs                            | 6.85±0.02μs                   |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 1, 1, 'd') |
| +        | 5.27±0.03μs                            | 6.84±0.01μs                   |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 1, 1, 'e') |
| +        | 5.26±0.01μs                            | 6.83±0.01μs                   |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 1, 2, 'e') |
| +        | 5.28±0.01μs                            | 6.84±0μs                      |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 1, 1, 'd')      |
| +        | 5.27±0.01μs                            | 6.86±0.02μs                   |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 1, 1, 'e')      |
| +        | 5.26±0.01μs                            | 6.85±0μs                      |    1.3  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 1, 2, 'e')      |
| +        | 5.29±0.01μs                            | 6.85±0.01μs                   |    1.29 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 4, 1, 'e') |
| +        | 5.29±0μs                               | 6.85±0μs                      |    1.29 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 4, 2, 'e') |
| +        | 5.29±0.02μs                            | 6.83±0.01μs                   |    1.29 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 4, 1, 'e') |
| +        | 5.29±0.02μs                            | 6.85±0.02μs                   |    1.29 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 4, 2, 'e') |
| +        | 5.30±0.01μs                            | 6.84±0.01μs                   |    1.29 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 4, 2, 'e')      |
| +        | 4.66±0.03μs                            | 6.02±0.2μs                    |    1.29 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'trunc'>, 4, 1, 'd')         |
| +        | 4.61±0μs                               | 5.85±0μs                      |    1.27 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'absolute'>, 4, 1, 'd')      |
| +        | 4.64±0.01μs                            | 5.83±0.01μs                   |    1.26 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'ceil'>, 4, 1, 'd')          |
| +        | 4.63±0.02μs                            | 5.86±0.01μs                   |    1.26 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'square'>, 4, 1, 'd')        |
| +        | 4.66±0.01μs                            | 5.83±0.01μs                   |    1.25 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'floor'>, 4, 1, 'd')         |
| +        | 4.67±0.02μs                            | 5.83±0.01μs                   |    1.25 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rint'>, 4, 1, 'd')          |
| +        | 4.35±0.02μs                            | 5.29±0.01μs                   |    1.22 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'ceil'>, 1, 2, 'f')          |
| +        | 4.36±0μs                               | 5.34±0.01μs                   |    1.22 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'floor'>, 1, 2, 'f')         |
| +        | 5.64±0.01μs                            | 6.86±0μs                      |    1.22 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 1, 2, 'd')      |
| +        | 4.35±0μs                               | 5.28±0.01μs                   |    1.22 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sqrt'>, 1, 2, 'f')          |
| +        | 4.36±0μs                               | 5.26±0μs                      |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'absolute'>, 1, 2, 'f')      |
| +        | 5.66±0.01μs                            | 6.85±0.01μs                   |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 1, 2, 'd') |
| +        | 4.36±0μs                               | 5.27±0μs                      |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rint'>, 1, 2, 'f')          |
| +        | 4.35±0.01μs                            | 5.25±0.01μs                   |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'square'>, 1, 2, 'f')        |
| +        | 4.36±0.01μs                            | 5.27±0μs                      |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'trunc'>, 1, 2, 'f')         |
| +        | 5.70±0.02μs                            | 6.85±0.01μs                   |    1.2  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 1, 2, 'd') |
| +        | 4.40±0.01μs                            | 5.30±0μs                      |    1.2  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'reciprocal'>, 1, 2, 'f')    |
| +        | 16.2±0.02μs                            | 19.2±0μs                      |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 1, 1, 'd')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 1, 1, 'f')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 1, 2, 'd')       |
| +        | 16.1±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 1, 2, 'f')       |
| +        | 16.1±0.02μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 4, 1, 'd')       |
| +        | 16.1±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 4, 1, 'f')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 4, 2, 'd')       |
| +        | 16.2±0.01μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'deg2rad'>, 4, 2, 'f')       |
| +        | 16.2±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 1, 1, 'd')       |
| +        | 16.2±0.07μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 1, 1, 'f')       |
| +        | 16.2±0.03μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 1, 2, 'd')       |
| +        | 16.2±0.04μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 1, 2, 'f')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 4, 1, 'd')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 4, 1, 'f')       |
| +        | 16.2±0.02μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 4, 2, 'd')       |
| +        | 16.2±0.05μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'degrees'>, 4, 2, 'f')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 1, 1, 'd')          |
| +        | 16.2±0.03μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 1, 1, 'f')          |
| +        | 16.1±0.02μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 1, 2, 'd')          |
| +        | 16.1±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 1, 2, 'f')          |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 4, 1, 'd')          |
| +        | 16.2±0.02μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 4, 1, 'f')          |
| +        | 16.2±0.01μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 4, 2, 'd')          |
| +        | 16.2±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'fabs'>, 4, 2, 'f')          |
| +        | 16.1±0.01μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 1, 1, 'd')       |
| +        | 16.1±0.01μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 1, 1, 'f')       |
| +        | 16.2±0.03μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 1, 2, 'd')       |
| +        | 16.2±0.01μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 1, 2, 'f')       |
| +        | 16.2±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 4, 1, 'd')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 4, 1, 'f')       |
| +        | 16.2±0.01μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 4, 2, 'd')       |
| +        | 16.2±0.02μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'rad2deg'>, 4, 2, 'f')       |
| +        | 16.2±0.03μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 1, 1, 'd')       |
| +        | 16.2±0.02μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 1, 1, 'f')       |
| +        | 16.2±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 1, 2, 'd')       |
| +        | 16.1±0μs                               | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 1, 2, 'f')       |
| +        | 16.2±0.03μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 4, 1, 'f')       |
| +        | 16.2±0.01μs                            | 19.2±0.02μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 4, 2, 'd')       |
| +        | 16.2±0.01μs                            | 19.2±0.01μs                   |    1.19 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 4, 2, 'f')       |
| +        | 5.80±0.01μs                            | 6.86±0.01μs                   |    1.18 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 4, 1, 'd') |
| +        | 5.81±0.01μs                            | 6.85±0.01μs                   |    1.18 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (1), 4, 1, 'd') |
| +        | 5.80±0.01μs                            | 6.87±0.01μs                   |    1.18 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 4, 1, 'd')      |
| +        | 16.4±0.1μs                             | 19.3±0.03μs                   |    1.17 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'radians'>, 4, 1, 'd')       |
| +        | 4.39±0.02μs                            | 4.88±0.4μs                    |    1.11 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'log10'>, 1, 1, 'e')         |
| +        | 59.6±0.3μs                             | 65.3±0.08μs                   |    1.1  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'reciprocal'>, 1, 1, 'e')    |
| +        | 59.5±0.03μs                            | 65.3±0.06μs                   |    1.1  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'reciprocal'>, 1, 2, 'e')    |
| +        | 59.6±0.02μs                            | 65.5±0.1μs                    |    1.1  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'reciprocal'>, 4, 2, 'e')    |
| +        | 59.6±0.09μs                            | 65.3±0.04μs                   |    1.09 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'reciprocal'>, 4, 1, 'e')    |
| +        | 7.00±0.03μs                            | 7.53±0.5μs                    |    1.08 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 1, 1, 'e')           |
| +        | 5.38±0.02μs                            | 5.69±0.02μs                   |    1.06 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'ceil'>, 1, 2, 'd')          |
| +        | 5.36±0μs                               | 5.66±0.01μs                   |    1.06 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'floor'>, 1, 2, 'd')         |
| +        | 7.85±0.03μs                            | 8.34±0.4μs                    |    1.06 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'log10'>, 1, 2, 'f')         |
| +        | 6.54±0.01μs                            | 6.94±0.09μs                   |    1.06 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'positive'>, 4, 2, 'd')      |
| +        | 5.36±0.01μs                            | 5.68±0.02μs                   |    1.06 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'trunc'>, 1, 2, 'd')         |
| +        | 6.51±0.03μs                            | 6.87±0μs                      |    1.05 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'conjugate'> (0), 4, 2, 'd') |
…